requirements.txt
asgiref==3.3.1
Django==3.1.5
djangorestframework==3.12.2
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pytesseract==0.3.7
pytz==2020.5
sqlparse==0.4.1
